requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
openai>=1.51.0
anthropic>=0.34.0
google-generativeai>=0.7.2
langdetect==1.0.9

